Make the ring any size: canonical water, a local site frame, distance-gridded meshes
The GPU only ever simulates canonical water (spacing 1, kernel 2, mass 1000,
sixty steps per water-second); a Resolution maps its metres by the spacing and
its seconds by Froude scaling, so water flows identically at every ring size
and never leaves single precision. Particles live in the drum's rotating frame
with static walls, the shader supplying centrifugal, Euler and Coriolis terms.

Bodies live in a local frame at a site on the wall, re-sited as the avatar
moves, with geometry that never forms the radius squared. The sky turns
instead of the drum. Meshes are gridded round the ring, along the axis and
across the caps with steps growing from the site, so no triangle is large
next to the viewer at any size; the settled avatar weighs one g on rings from
a kilometre to a light year, and a frame of any of them draws the ground.

The test harness makes one GPU device per process: creating and tearing down
a device per app faulted the NVIDIA driver (Xid 31/32/13) about once in eight
suite runs, which was the flake. Captured frames have their row padding
stripped.

+++ game/tests/persistence.rs
@@ -61,7 +61,7 @@ fn snapshot_round_trips_through_json() {
         );
     }
     assert_eq!(restored.fluid.len(), 40 * 7);
-    assert_eq!(restored.fluid[0], particles[0].position[0]);
+    assert_eq!(restored.fluid[0], particles[0].position[0] as f32);
 }
 
 #[test]
